♻️ [refactor]: #452 Phase 3a: lexer_advance.rs を分離
advance の 4 テストを `src/lexer/tests/lexer_advance.rs` に分離。
Phase 3 ヘッダコメントは Phase 3b までは tests.rs に残す。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_advance.rs: 36 行 (< 800)

Refs #452
